1.20.1-4.1.0:
Additions
-Added sounds to reviving, fallen state, when fully revived, and when calling for help
-Added Call for Help button
-You can now stop Command/Regular chat messages from displaying through the config

Changes
-Decreased distance you can see a fallen player when they are not calling for help
-Now a revive chat message is sent when you revive someone normally

Fixes
-Mobs should now fully ignore you while in the fallen state
-No longer displays the wrong number while dying if your world is old
-Food level now updates on CLIENT when used to revive someone

+++ src/main/java/invoker54/reviveme/mixin/ForgeHooksMixin.java
@@ -50,9 +50,9 @@ public class ForgeHooksMixin {
         if (!targCap.hasEnough(player)) return;
 
         //Now add the player to the targets fallencapability and vice versa.
-        targCap.setProgress((int) player.level().getGameTime(), ReviveMeConfig.reviveTime);
+        targCap.setProgress(player.level().getGameTime(), ReviveMeConfig.reviveTime);
         targCap.setOtherPlayer(player.getUUID());
-        myCap.setProgress((int) player.level().getGameTime(), ReviveMeConfig.reviveTime);
+        myCap.setProgress(player.level().getGameTime(), ReviveMeConfig.reviveTime);
         myCap.setOtherPlayer(targPlayer.getUUID());
 
         //Make sure the fallen client has this data too
